requirements.txt
google-genai>=1.15
litellm>=1.67
numpy>=1.25
openai>=1.3
pandas>=2.0
platformdirs>=3.10
requests>=2.31
scikit-learn>=1.3
tenacity>=8.2
textgrad==0.1.8
tiktoken>=0.5
torch>=2.2
